Fix issue #354 test for 32-bit platforms
The vendored issue #354 test called `llvmcall_354(10)`, but a bare integer
literal is `Int32` on 32-bit platforms and so does not match the method's
`Int64` signature, giving a `MethodError`. Pass `Int64(10)` explicitly.

The failure was masked on Windows CI until PyCall, whose build failed before
any test ran, was dropped.

+++ test/interpret.jl
@@ -258,7 +258,9 @@ function llvmcall_354(x::Int64)
     Base.llvmcall("ret i64 %0", Int64, Tuple{Int64}, x + 1)
 end
 push!(JuliaInterpreter.compiled_methods, which(llvmcall_354, Tuple{Int64}))
-@test @interpret(llvmcall_354(10)) === Int64(11)
+# `Int64(10)`, not `10`: a bare literal is `Int32` on 32-bit platforms and would
+# not match the `Int64` signature.
+@test @interpret(llvmcall_354(Int64(10))) === Int64(11)
 
 # "correct" line numbers
 defline = @__LINE__() + 1
